# specs/001-ugc-video-generator/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Viral4Creators - UGC Video Cloner API
  description: |
    API for generating advertisement videos based on UGC (User Generated Content) analysis.
    
    ## Workflow
    1. Create session
    2. Upload source video to S3
    3. Trigger video analysis (Gemini AI)
    4. Submit product information
    5. Upload product image to S3
    6. Generate video prompt (GPT-5)
    7. Generate video (Sora 2)
    8. Download generated video
    
    This is a stateless POC - session data is held in memory only.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /sessions/{sessionId}/video/upload-url:
    post:
      tags:
        - Video
      summary: Get presigned upload URL for video
      description: Generate a presigned S3 POST URL for uploading source video directly from browser
      operationId: getVideoUploadUrl
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadVideoRequest'
      responses:
        '200':
          description: Presigned URL generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadVideoResponse'
        '400':
          description: Invalid request (file too large, invalid format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /sessions/{sessionId}/prompt:
    post:
      tags:
        - Prompt
      summary: Generate video prompt
      description: Generate text-to-video prompt using GPT-5 based on analysis and product info
      operationId: generatePrompt
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Prompt generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratePromptResponse'
        '400':
          description: Missing required data (analysis or product info)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Prompt
      summary: Update prompt text
      description: Save user's edited version of the generated prompt
      operationId: updatePrompt
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePromptRequest'
      responses:
        '200':
          description: Prompt updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatePromptResponse'
        '400':
          description: Invalid input (too long, empty)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{sessionId}/prompt/approve:
    post:
      tags:
        - Prompt
      summary: Approve prompt for video generation
      description: Mark prompt as approved and ready for video generation
      operationId: approvePrompt
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Prompt approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovePromptResponse'
        '400':
          description: Prompt not generated yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      description: Unique session identifier (UUID)
      schema:
        type: string
        format: uuid
        example: 123e4567-e89b-12d3-a456-426614174000
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - meta
      properties:
        success:
          type: boolean
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      required:
        - timestamp
        - requestId
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - error
          properties:
            error:
              $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VIDEO_TOO_LARGE
        message:
          type: string
          example: Video file exceeds maximum size of 100MB

    # Session schemas
...
    UploadVideoRequest:
      type: object
      required:
        - fileName
        - fileSize
        - mimeType
      properties:
        fileName:
          type: string
          example: my-video.mp4
        fileSize:
          type: integer
          format: int64
          maximum: 104857600
          description: File size in bytes (max 100MB)
          example: 52428800
        mimeType:
          type: string
          enum: [video/mp4, video/quicktime, video/x-msvideo]
          example: video/mp4
    UploadVideoResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - uploadUrl
                - uploadFields
                - s3Key
              properties:
                uploadUrl:
                  type: string
                  format: uri
                  description: S3 presigned POST URL
                uploadFields:
                  type: object
                  additionalProperties:
                    type: string
                  description: Form fields to include in POST request
                s3Key:
                  type: string
                  description: S3 key where file will be stored
...
    GeneratePromptResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/GenerationPrompt'
    UpdatePromptRequest:
      type: object
      required:
        - editedText
      properties:
        editedText:
          type: string
          minLength: 1
          maxLength: 500
    UpdatePromptResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/GenerationPrompt'
    ApprovePromptResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/GenerationPrompt'
    GenerationPrompt:
      type: object
      required:
        - promptId
        - generatedText
        - finalText
        - characterCount
        - generatedAt
        - moderationStatus
      properties:
        promptId:
          type: string
          format: uuid
        generatedText:
          type: string
          description: AI-generated prompt
        userEditedText:
          type: string
          description: User's edited version
        finalText:
          type: string
          description: Final approved text
        characterCount:
          type: integer
        generatedAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
        moderationStatus:
          type: string
          enum: [pending, approved, flagged, bypassed]
        moderationFlags:
          type: array
          items:
            type: string

    # Generation schemas
...
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (optional for POC)

# Security is optional for POC but documented for future use
security: []